+++ ikuzo/domain/organization.go
@@ -28,7 +28,7 @@ type orgIDKey struct{}
 var (
 	ErrIDTooLong          = errors.New("identifier is too long")
 	ErrIDNotLowercase     = errors.New("uppercase not allowed in identifier")
-	ErrIDInvalidCharacter = errors.New("only letters are allowed in organization")
+	ErrIDInvalidCharacter = errors.New("only letters and numbers are allowed in organization")
 	ErrIDCannotBeEmpty    = errors.New("empty string is not a valid identifier")
 	ErrIDExists           = errors.New("identifier already exists")
 	ErrOrgNotFound        = errors.New("organization not found")
@@ -117,9 +117,11 @@ func (id OrganizationID) Valid() error {
 			continue
 		}
 
-		if !unicode.IsLetter(r) {
-			return ErrIDInvalidCharacter
+		if unicode.IsLetter(r) || unicode.IsNumber(r) {
+			continue
 		}
+
+		return ErrIDInvalidCharacter
 	}
 
 	return nil