+++ CHANGELOG.md
@@ -4,16 +4,20 @@
 
 - history of changes: see https://github.com/delving/hub3/compare/v0.2.0...master
 
-### Added 
+### Added
 
+- ClI subcommand 'bulk' to index bulk-requests that are serialized to disk [[GH-88]](https://github.com/delving/hub3/pull/88)
 - Config option to store records generated from METS files in dedicated index [[GH-83]](https://github.com/delving/hub3/pull/83)
 
 ## Changed 
 - Allow for changes to uploaded EAD file before storing it [[GH-90]](https://github.com/delving/hub3/pull/90)
 -  Refactored EAD search overview queries into separate filter and collapse queries [[GH-91]](https://github.com/delving/hub3/pull/91)
+- Added support for NDE Dataset Register API [[GH-92]](https://github.com/delving/hub3/pull/92)
+
 
 ### Fixed
-
+ 
+- Explicit initialisation of config for CLI subcommands instead of global on init [[GH-88]](https://github.com/delving/hub3/pull/88)
 - Prevent duplicate files while processing METS files [[GH83]](https://github.com/delving/hub3/pull/89)
 
 ## v0.2.0 (2021-03-10)